add :tractor: pallet and requirements.txt reminder
This will add a `sites.json` file to a data folder next to the pallet.
Since we are [generating reports](#110) and [site popups](#111) from
this data, I just wanted to get it stored. We can modify the pallet at a
later time to create the reports as well.

i doubt :tractor: is in pip but it's a good reminder of the
python
dependencies none the less.

closes #112